requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
supabase>=2.6.0
openai>=1.43.0
httpx>=0.27.0
python-dotenv>=1.0.1
